~ db - Copier les rum dans une db

Copier les rum, les actes et les diagnostics des rums dans une db

db_mco_in(con, p, remove = T, zip = T, indexes = list(), ...)

Arguments

con

la connexion a la base de donnees (src_..)

p

le noyau pmeasyr

remove

a TRUE, les tables precedentes rum de l'annee sont effacees avant

zip

a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db

indexes

index a ajouter a la table dans la base (voir copy_to)

Value

nothing

Details

Les tables sont importées dans R puis copiées dans la db La table diag est créée et la durée des rum est calculée (DUREESEJPART)

Examples

if (FALSE) {
purrr::quietly(db_mco_in)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_mco_in)(con, p, annee = 2015) -> statuts ; gc(); #..
}