~ MCO - Import des Med

Import des fichiers MED In ou Out.

imed_mco(finess, annee, mois, path, lib = T, tolower_names = F, typmed = c('out', 'in'))

Arguments

finess
Finess du Out a importer : dans le nom du fichier

annee
Annee PMSI (nb) des donnees sur 4 caracteres (2016)

mois
Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)

path
Localisation du fichier de donnees

typmed
Type de donnees In / Out

lib
Ajout des libelles de colonnes aux tables, par defaut a TRUE ; necessite le package sjlabelled

tolower_names
a TRUE les noms de colonnes sont tous en minuscules

~...
parametres supplementaires a passer dans la fonction read_fwf, par exemple n_max = 1e3 pour lire les 1000 premieres lignes, progress = F, skip = 1e3

Value

Une table (data.frame, tibble) contenant les médicaments In ou Out (T2A, ATU et thrombo selon l'existence des fichiers : si le fichier n'existe pas, pas de donnée importée). Pour discriminer le type de prestation, la colonne TYPEPREST donne l'information : T2A 06 - ATU 09 - THROMBO 10

Details

Formats depuis 2011 pris en charge

See also

Examples

# NOT RUN {
   imed_mco('750712184',2015,12,'~/Documents/data/mco') -> med_out15
   imed_mco('750712184',2015,12,'~/Documents/data/mco', typmed = "in") -> med_in15
# }