~ MCO - Import des PO

Imports des fichiers PO In / Out

ipo(finess, annee, mois, path, typpo = c("out", "in"), lib = T, ...)

Arguments

finess
Finess du Out a importer : dans le nom du fichier

annee
Annee PMSI (nb) des donnees sur 4 caracteres (2016)

mois
Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)

path
Localisation du fichier de donnees

typpo
Type de donnees In / Out

lib
Ajout des libelles de colonnes aux tables, par defaut a TRUE ; necessite le package sjlabelled

~...
parametres supplementaires a passer dans la fonction read_fwf, par exemple n_max = 1e3 pour lire les 1000 premieres lignes, progress = F, skip = 1e3

Value

Une table (data.frame, tibble) contenant les prélèvements d'organes In ou Out.

Details

Formats depuis 2011 pris en charge

See also

irum, irsa, utiliser un noyau de parametres avec noyau_pmeasyr

Examples

not_run({
     po <- ipo('750712184',2015,12,"~/Documents/data/mco")
})