Découpe un seul nom de fichier

Découpe un seul nom de fichier

parse_nom_fichier(nom_fichier, format_date_archive = "%d%m%Y%H%M%S")

Arguments

nom_fichier
Le nom du fichier à découper. Chaine de caractère de longueur 1.

format_date_archive
Format de date d'horodatage pour les fichiers archive avec la notation de base::strptime().

Value

Une liste ou une table avec les informations extraites du nom de fichier :

nom_fichier Le nom de fichier passé en arguement

finess

annee

mois

horodatage_production L'horodatage de production pour les fichiers in et out au format POSIXlt Si il y a plusieurs fichiers, la liste peut être facilement transformée en table avec dplyr::bind_rows().

type Type de fichier : in, out, rss...

See also