Selectionne une archive in/out

Selectionne une archive in/out

selectionne_archive(finess, mois, annee, dossier_archives,
  type_archive = "in", recent = TRUE)

Arguments

finess
Finess de l'archive.

mois
Mois de l'archive.

annee
Année de l'archive.

dossier_archives
Chemin vers le dossier contenant les archives. Vecteur de caractère de longueur 1.

type_archive
Type d'archive : in (par défaut) ou out

recent
par défaut TRUE, l'archive la plus recente sera utilisee, sinon propose à l'utilisateur de choisir quelle archive dezipper

Value

Une liste ou une table avec les informations extraites du nom de fichier :

nom_fichier Le nom de fichier passé en arguement

finess

annee

mois

horodatage_production L'horodatage de production pour les fichiers in et out au format POSIXlt Si il y a plusieurs fichiers, la liste peut être facilement transformée en table avec dplyr::bind_rows().

type Type de fichier : in, out, rss...

See also