~ *.zip - Dezippe des fichiers de l'archive PMSI

Dezipper une archive PMSI au besoin

adezip(finess, annee, mois, path, liste, pathto = "", type, recent = T)

Arguments

finess
Finess du fichier a dezipper

annee
Annee du fichier

mois
Mois du fichier

path
Chemin d'acces au fichier

liste
des fichiers à dezipper ex: ano, rss, rsa, dmi, ... ; si liste = "", dezippe la totalite de l'archive

type
Type de l'archive : in / out

recent
par defaut a T, l'archive la plus recente sera utilisee, sinon propose a l'utilisateur de choisir quelle archive dezipper

pathto
par defaut a "", dezipper la ou est l'archive, sinon preciser le chemin ou dezipper les fichiers (ailleurs)

See also

adezip2, astat, adelete, utiliser un noyau de parametres avec noyau_pmeasyr

Examples

# NOT RUN {
     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = 'med',
            pathto = "~/Exemple",
            type = "out")

     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = c('med','rapss', 'ano'),
            pathto = "~/Exemple",
            type = "in")

     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = c('rss', 'ano'),
            pathto = "~/Exemple",
            type = "in",
            recent = F)
# }