~ *.zip - Dezippe des fichiers de l'archive PMSI, avec en parametre le nom de l'archive

Alternative à la fonction adezip(), si on connait précisement l'archive que l'on veut utiliser.

adezip2(path, file, liste = "", pathto = "")

Arguments

path
Chemin d'accès au répertoire contenant l'archive à décompresser

file
Nom de l'archive zip (ex: 750712184.2016.2.05042016093044.in.zip)

liste
Vecteur de caractère avec le type de fichiers à dézipper (ex: ano, rss, rsa, dmi, ...). Par défaut, liste = "" dezippe la totalite de l'archive.

pathto
Par defaut la même valeur que path, dézipper dans le même répertoire que l'archive, sinon préciser le chemin ou dezipper les fichiers dans le répertoire indiqué par pathto.

Details

adezip2 est un simple wrapper autour de la fonction adzip.default. Cette fonction est dépréciée. En utilisant le paramètre nom_archive avec la fonciton adezip() on obtient le même résultat.

See also

adezip, astat, adelete

Examples

# Chemin vers un dossier temporaire
tmp_dir <- tempdir()

# Chemin vers un dossier contenant des archives simulées
dossier_archives <- system.file("extdata", "test_data", "test_adezip", package = "pmeasyr")

# Décompresser en fonction du finess, année et mois du 
# fichier med d'une archive out
 adezip2(path = dossier_archives,
         file = "123456789.2017.7.21082017091715.out.zip",
         liste = 'med',
         pathto = tmp_dir)

#> 
#> Dézippage de l'archive 123456789.2017.7.21082017091715.out.zip
#> Taille    : 0.0072 Mo
#> Type      : out
#> Finess    : 123456789
#> Période   : 2017 M7
#> Date prod : 2017-08-21 09:17:15
#> Fichiers  : med

dir(tmp_dir, pattern = "2017\\.7.*med")

#> [1] "123456789.2017.7.med"