~ *.zip - Dezippe des fichiers de l'archive PMSI en provenance de l'Intranet AP-HP, avec en parametre le nom de l'archive

Version de la fonction adezip2 pour des archives au format Intranet du DIM Siege de l'AP-HP, http://dime.aphp.fr/.

adezip3(finess, path, file, liste = "", pathto = "")

Arguments

finess

Finess du fichier a dezipper

path

Chemin d'acces au fichier

file

Nom de l'archive zip (ex: MCO_IN_00000_201603.zip)

liste

des fichiers a dezipper ex: ano, rss, rsa, dmi, ... ; si liste = "", dezippe la totalite de l'archive

pathto

Chemin ou deposer les fichiers dezippes, par defaut à "", les fichiers sont mis la ou se trouve l'archive

See also

Author

G. Pressiat

Examples

if (FALSE) { # \dontrun{
    # Fichier ano
    adezip3(path = '~/Downloads',
            file = 'MCO_IN_00000_201603.zip',
            liste = 'ano')

    # Totalité de l'archive
    adezip3(path = '~/Downloads',
            file = 'MCO_IN_00000_201603.zip',
            liste = '')
} # }