Decoupe un seul nom de fichier

Les noms de fichiers des archives PMSI ont toujours le même format, de la forme finess.annee_envois.mois_envois. suivi par horodatage pour les archives in et ou et enfin le type de fichier. Cette fonction permet d'extraire cette information.

parse_nom_fichier(nom_fichier, format_date_archive = "%d%m%Y%H%M%S")

Arguments

nom_fichier
Le nom du fichier à découper. Chaine de caractère de longueur 1.

format_date_archive
Format de date d'horodatage pour les fichiers archive avec la notation de base::strptime().

Value

Une liste ou une table avec les informations extraites du nom de fichier :

nom_fichier Le nom de fichier passé en arguement

finess

annee

mois

horodatage_production L'horodatage de production pour les fichiers in et out au format POSIXlt Si il y a plusieurs fichiers, la liste peut être facilement transformée en table avec dplyr::bind_rows().

type Type de fichier : in, out, rss...

See also

Examples

noms_de_fichiers <- c("671234567.2016.1.12032016140012.in.zip",
                      "671234567.2016.1.rum.txt")
 x <- parse_noms_fichiers(noms_de_fichiers)
dplyr::bind_rows(x)

#> # A tibble: 2 x 6
#>   nom_fichier                   finess   annee mois  horodatage_producti… type  
#>   <chr>                         <chr>    <chr> <chr> <dttm>               <chr> 
#> 1 671234567.2016.1.12032016140… 6712345… 2016  1     2016-03-12 14:00:12  in    
#> 2 671234567.2016.1.rum.txt      6712345… 2016  1     NA                   rum.t…