Extraire les informations d'un nom de fichier

Les noms de fichiers des archives PMSI ont toujours le même format, de la forme finess.annee_envois.mois_envois. suivi par horodatage pour les archives in et ou et enfin le type de fichier. Cette fonction permet d'extraire cette information.

parse_noms_fichiers(noms_fichiers, return_tibble = TRUE)

Arguments

noms_fichiers

Des noms de fichiers d'archive sous la forme d'un vecteur de caractères.

return_tible

Retourne une table si TRUE, sinon retourne une liste.

Value

Une liste ou une table avec les informations extraites du nom de fichier :

nom_fichier Le nom de fichier passé en arguement

finess

annee

mois

horodatage_production L'horodatage de production pour les fichiers in et out au format POSIXlt Si il y a plusieurs fichiers, la liste peut être facilement transformée en table avec dplyr::bind_rows().

type Type de fichier : in, out, rss...

Examples

if (FALSE) { # \dontrun{
noms_de_fichiers <- c("671234567.2016.1.12032016140012.in.zip",
                      "671234567.2016.1.rum.txt")
 x <- parse_noms_fichiers(noms_de_fichiers)
dplyr::bind_rows(x)
} # }