~ VVR - Attribuer les recettes GHS et suppléments sur des rsa

Reproduire la valorisation BR et coefficients géo/prudentiels du tableau RAV d'epmsi, à partir des tables résultant des fonctions vvr_rsa, vvr_ano_mco, et de tables contenant les fichcomp PO et DIAP, PIE et MO

vvr_ghs_supp(rsa, tarifs, supplements = NULL, ano = NULL,
  porg = dplyr::tibble(), diap = dplyr::tibble(),
  pie = dplyr::tibble(), mo = dplyr::tibble(), full = FALSE,
  cgeo = 1.07, prudent = NULL, bee = TRUE)

Arguments

rsa
Un tibble rsa partie fixe (créé avec vvr_rsa)

tarifs
Un tibble contenant une ligne par tarif GHS - année séquentielle des tarifs

supplements
Un tibble contenant une ligne par année et une colonne par tarif de supplément

ano
Un tibble, facultatif si `r bee = TRUE`, créé avec vvr_ano_mco

porg
Un tibble contenant les prélévements d'organes du out (importés avec ipo)

diap
Un tibble contenant les dialyses péritonéales du out (importés avec idiap)

pie
Un tibble contenant les prestations inter-établissements du out (importés avec ipie)

mo
Un tibble contenant les molécules onéreuses du out (importés avec imed_mco)

full
Booléen, à `r TRUE` toutes les variables intermédiaires de valo sont gardées

cgeo
Coefficient géographique, par défaut celui de l'Île-de-France (1.07)

prudent
coefficient prudentiel, par défaut à `r NULL`, le coefficient prudentiel est appliqué par année séquentielles des tarifs

bee
par défaut à `r TRUE`, seule la valorisation de GHS de base + extrême haut - extrême bas est calculée

Value

Un tibble contenant les différentes rubriques de valorisation, une ligne par clé rsa

Details

Cette fonction ne tient pas compte de la rubrique "Minoration forfaitaire liste en sus, car elle a été supprimée en 2018

See also

Examples

# NOT RUN {
# Récupérer les tarifs GHS et des suppléments (ex-DGF) : 
tarifs      <- nomensland::get_table('tarifs_mco_ghs') %>% distinct(ghs, anseqta, .keep_all = TRUE)
supplements <- nomensland::get_table('tarifs_mco_supplements') %>% mutate_if(is.numeric, tidyr::replace_na, 0)

# Recette GHS de base et suppléments EXB, EXH
vvr_ghs_supp(rsa = vrsa, tarifs = tarifs)

# Recette GHS de base et suppléments EXB, EXH, et des suppléments
vvr_ghs_supp(vrsa, tarifs, supplements, vano, ipo(p), idiap(p), ipie(p), imed_mco(p), bee = FALSE)
# }